gpu/capture: lift the byte cursors and per-field codecs into capture_codecs.hpp
The Writer/Reader cursors and the 20 write_*/read_* field codecs are the vocabulary the
serialize and deserialize entry points share, so they have to be visible from both before
those two can live in separate files. 22 regions change linkage; 20 gain `inline`.

capture_codecs.hpp 422 lines; gpu_capture_serialize.cpp -> 2481.
